Feature/abutting kata overlay (#571)
* New feature: config option show-kata-estimate-large displays estimate rectangles large enough that they abut, with more transparency, and _behind_ the stones, which some users -- including me -- may prefer.
* Separate on/off and mode-cycling. Provide both 'smart' modes. More explicit mode names. Mostly-fix race condition issue.

+++ src/main/java/featurecat/lizzie/gui/MainFrame.java
@@ -361,7 +361,7 @@ public abstract class MainFrame extends JFrame {
 
   public abstract void noEstimateByZen(boolean byToolBar);
 
-  public abstract void drawEstimateRectZen(ArrayList<Integer> esitmateArray);
+  public abstract void drawEstimateRectZen(ArrayList<Double> esitmateArray);
 
   public void saveImage() {};
